+++ code/network/multiui.cpp
@@ -4647,7 +4647,7 @@ void multi_create_list_do()
 	int y_start = Mc_list_coords[gr_screen.res][MC_Y_COORD];		
 
 	start_index = multi_create_select_to_index(Multi_create_list_start);
-	stop_index = (Multi_create_list_mode == MULTI_CREATE_SHOW_MISSIONS) ? Multi_create_mission_list.size() : Multi_create_campaign_list.size();
+	stop_index = (int)((Multi_create_list_mode == MULTI_CREATE_SHOW_MISSIONS) ? Multi_create_mission_list.size() : Multi_create_campaign_list.size());
 
 	for (idx = start_index; idx < stop_index; idx++) {
 		multi_create_info *mcip;
@@ -4864,7 +4864,7 @@ void multi_create_list_blit_icons(int list_index, int y_start)
 	int max_index;
 
 	// get a pointer to the list item
-	max_index = (Multi_create_list_mode == MULTI_CREATE_SHOW_MISSIONS) ? Multi_create_mission_list.size() - 1 : Multi_create_campaign_list.size() - 1;
+	max_index = (int)((Multi_create_list_mode == MULTI_CREATE_SHOW_MISSIONS) ? Multi_create_mission_list.size() - 1 : Multi_create_campaign_list.size() - 1);
 
 	if ( (list_index < 0) || (list_index > max_index) ) {
 		return;
